RFC: Replace libuuid and compat with internal implementation
We use a very small subset of libuuid features to transform UUIDs
between their string and binary representation. Plus we have a lot of
compat code for different platforms with some unspecified default
behavior regarding the use of upper/lower case.

Drop the dependency on libuuid and replace it with a minimal internal
implementation that respects RFC4122, inspired by the MIT licensed
version from the Azure C shared utility library.

This is an RFC for the concept, a final version would have proper logging
and better memory allocation policies, the current version builds and
pass the tests.

Change-Id: Ic170ce26ade23d177195cad117bd0fab590b328e
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>