Add .pre-commit-config.yaml
This can help catch problems a bit earlier than in the CI.

To use it, a developer would run `pre-commit` in the source repository.
This installs git commit hooks that will run the checks described by
`.pre-commit-config.yaml` on the modified files.

It is possible to run the checks on all files with:

    $ pre-commit run --all-files
    black....................................................................Passed
    flake8...................................................................Passed
    isort....................................................................Passed

Change-Id: I695f0b0f2fea3adadae132ceb708928db3217e0c
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13202
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>
Tested-by: jenkins <jenkins@lttng.org>