pyproject.toml: pyright: ignore reportPrivateUsage
Our code does not follow the convention of treating identifiers starting
with an underscore as private.  For instance, the `bt2` module exposes
types like `_EventMessageConst`.  Even though the user is not expected
to create objects of that type, they can still use the type in a type
check like:

    if type(msg) is bt2._EventMessageConst:

That produces a `reportPrivateUsage` in pyright / pylance.  Disable that
check to get rid of a lot of false positives.

Change-Id: I78a19653aafb923d3e429b9b321cb1f87372c83e
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13307
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>
Tested-by: jenkins <jenkins@lttng.org>