tests: factor out discarded packet regex in test-packet-seq-num.sh
The regex used in `test_no_lost()` is outdated, which means that if
Babeltrace were to emit an unexpected discarded packets message, the
test would not catch it.  Fix it by factoring the regex used to look for
discarded packets messages.  At least now, we can have the confidence
that we're looking for the right thing, otherwise the `test_lost()`
tests wouldn't pass.

In general, it's fragile to check for the absence of something, since a
typo or change in what to look can easily make the test uneffective.  I
initially re-wrote this test in Python (the test is in the `cli`
directory, but its intent is more to validate the source's behavior than
to validate the CLI's behavior), but this simple change seems good
enough for now.

Change-Id: I9f8c30a567fa6f5036cadcc4691c3bd6f953ca3d
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13309
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>